Move EVM account generation code to common crate
