+++ crates/sp-domains/src/test_ethereum.rs
@@ -3,10 +3,46 @@
 use crate::test_ethereum_tx::{
     AccountInfo, EIP1559UnsignedTransaction, EIP2930UnsignedTransaction, LegacyUnsignedTransaction,
 };
+use crate::{EthereumAccountId, PermissionedActionAllowedBy};
 use ethereum::TransactionV2 as Transaction;
 use frame_support::pallet_prelude::DispatchClass;
+use hexlit::hex;
 use pallet_evm::GasWeightMapping;
 use sp_core::{Get, U256};
+
+/// The kind of account list to generate.
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+pub enum EvmAccountList {
+    Anyone,
+    NoOne,
+    One,
+    Multiple,
+}
+
+/// Generate the supplied kind of account list.
+pub fn generate_evm_account_list(
+    account_infos: &[AccountInfo],
+    account_list_type: EvmAccountList,
+) -> PermissionedActionAllowedBy<EthereumAccountId> {
+    // The signer of pallet-evm transactions is the EVM domain in some tests, so we also add it to
+    // the account lists.
+    let evm_domain_account = hex!("e04cc55ebee1cbce552f250e85c57b70b2e2625b");
+
+    match account_list_type {
+        EvmAccountList::Anyone => PermissionedActionAllowedBy::Anyone,
+        EvmAccountList::NoOne => PermissionedActionAllowedBy::Accounts(Vec::new()),
+        EvmAccountList::One => PermissionedActionAllowedBy::Accounts(vec![
+            EthereumAccountId::from(evm_domain_account),
+            EthereumAccountId::from(account_infos[0].address),
+        ]),
+        EvmAccountList::Multiple => PermissionedActionAllowedBy::Accounts(vec![
+            EthereumAccountId::from(evm_domain_account),
+            EthereumAccountId::from(account_infos[0].address),
+            EthereumAccountId::from(account_infos[1].address),
+            EthereumAccountId::from(account_infos[2].address),
+        ]),
+    }
+}
 
 pub fn max_extrinsic_gas<TestRuntime: frame_system::Config + pallet_evm::Config>(
     multiplier: u64,
